refactor: utiliza record para o input do search de ações
